Fix a crash when a package file cannot be open
Sometimes when downloading a package some symbolic links files cannot be
open on Windows when Nimble is run as administrator. For this reason,
add exception handling to the `open` procedure to avoid the crash and
display a warning message that the file content will not be count in the
calculation of the package's checksum.

Related to nim-lang/nimble#127